[*] change init db

+++ services/sqlLiteServices/dbInit.js
@@ -3,31 +3,36 @@ import HttpError from "../../helpers/HttpError.js";
 import { serviceLogger } from "../../config/logConfig.js";
 import { parseXlsAndSaveToDb } from "../fileParserServices/parseXlsAndSaveToDb.js";
 
+const DB_INIT = process.env.DB_INIT || "false";
+
+console.log("DB_INIT: ", DB_INIT);
+serviceLogger.info(`DB initialize ENABLED: ${DB_INIT}`);
+
 export const initializeDatabase = async () => {
   try {
     const db = new Database("database.sqlite");
 
     serviceLogger.info("Connected to SQLite database");
+    if (DB_INIT === "true") {
+      // Видалення існуючих таблиць, якщо вони вже є
+      const dropTypesSchemaQuery = `DROP TABLE IF EXISTS types;`;
+      db.exec(dropTypesSchemaQuery);
 
-    // Видалення існуючих таблиць, якщо вони вже є
-    const dropTypesSchemaQuery = `DROP TABLE IF EXISTS types;`;
-    db.exec(dropTypesSchemaQuery);
+      const dropKatotgSchemaQuery = `DROP TABLE IF EXISTS katotg;`;
+      db.exec(dropKatotgSchemaQuery);
 
-    const dropKatotgSchemaQuery = `DROP TABLE IF EXISTS katotg;`;
-    db.exec(dropKatotgSchemaQuery);
-
-    // Створюємо таблицю types
-    const createTypesTableQuery = `
+      // Створюємо таблицю types
+      const createTypesTableQuery = `
       CREATE TABLE IF NOT EXISTS types (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         type TEXT NOT NULL,
         name TEXT NOT NULL
       );
     `;
-    db.exec(createTypesTableQuery);
+      db.exec(createTypesTableQuery);
 
-    // Створюємо таблицю katotg
-    const createKatotgTableQuery = `
+      // Створюємо таблицю katotg
+      const createKatotgTableQuery = `
       CREATE TABLE IF NOT EXISTS katotg (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         katotg TEXT NOT NULL,
@@ -36,11 +41,12 @@ export const initializeDatabase = async () => {
         dps_code TEXT NOT NULL
       );
     `;
-    db.exec(createKatotgTableQuery);
+      db.exec(createKatotgTableQuery);
 
-    await parseXlsAndSaveToDb(db);
+      await parseXlsAndSaveToDb(db);
 
-    serviceLogger.info("Database tables initialized");
+      serviceLogger.info("Database tables initialized");
+    }
     return db;
   } catch (error) {
     serviceLogger.error("Error details:", error);
